+++ library/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the library will be documented in this file.
 
 - Add `checkItemsAsync` action (pull request #856)
 - Add `graphemes`, `maxGraphemes`, `minGraphemes` and `notGraphemes` action (pull request #853)
+- Add `words`, `maxWords`, `minWords` and `notWords` action
 - Change types and implementation to support Standard Schema
 - Refactor `bytes`, `maxBytes`, `minBytes` and `notBytes` action